src/main/java/com/example/ticketBooking/service/impl/FeedbackServiceImpl.java
package com.example.ticketBooking.service.impl;

import com.example.ticketBooking.entity.Feedback;
import com.example.ticketBooking.repository.FeedbackRepository;
import com.example.ticketBooking.service.FeedbackService;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor

public class FeedbackServiceImpl implements FeedbackService {

    private FeedbackRepository feedbackRepository;

    @Override
    public List<Feedback> getAllFeedbacks() {
        return feedbackRepository.findAll();
    }

    @Override
    public void saveFeedback(Feedback feedback) {
        feedbackRepository.save(feedback);
    }
}
